backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.28.0
python-dotenv>=1.0.0
cryptography>=44.0.0
pillow>=11.0.0
numpy>=2.0.0
opencv-python-headless>=4.10.0
python-multipart>=0.0.18
